requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.6
python-docx==1.1.0
pydantic==2.4.2
pydantic-settings==2.0.3

# AI API clients
google-generativeai==0.3.0
anthropic==0.7.1
openai==1.3.5

# Database and storage
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx>=0.28.1
tenacity==8.2.3
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
